User w/Foreign key

+++ MyBlog/posts/views.py
@@ -1,6 +1,6 @@
 from urllib.parse import quote_plus
 from django.contrib import messages
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
@@ -10,9 +10,12 @@ from .forms import PostForm
 
 # Create your views here.
 def post_create(request):
+    if request.user.is_staff or request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, 'Successfully Created!')
         return HttpResponseRedirect(instance.get_absoulte_url())
@@ -53,6 +56,8 @@ def post_list(request):
 
 
 def post_update(request, id=None):
+    if request.user.is_staff or request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -70,6 +75,8 @@ def post_update(request, id=None):
 
 
 def post_delete(request, id=None):
+    if request.user.is_staff or request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     instance.delete()
     messages.success(request, 'Successfully deleted!')
